# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 1.0.0
  description: Personal Finance & Accounting App API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /settings:
    get:
      operationId: getSettings
      tags: [settings]
      summary: Get user settings
      security:
        - bearerAuth: []
      responses:
        "200":
          description: User settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserSettings"
    put:
      operationId: updateSettings
      tags: [settings]
      summary: Update user settings
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateSettingsRequest"
      responses:
        "200":
          description: Settings updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserSettings"

  # ─── ADMIN ───────────────────────────────────────────────────
  /admin/stats:
    get:
      operationId: getAdminStats
      tags: [admin]
      summary: Get admin dashboard statistics
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Admin statistics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminStats"
  /admin/revenue:
    get:
      operationId: getAdminRevenue
      tags: [admin]
      summary: Get revenue chart data
      security:
        - bearerAuth: []
      parameters:
        - name: period
          in: query
          schema:
            type: string
            enum: [week, month, quarter, year]
            default: month
      responses:
        "200":
          description: Revenue data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RevenueData"
  /admin/users:
    get:
      operationId: getAdminUsers
      tags: [admin]
      summary: Get all users (admin)
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: number
            default: 1
        - name: limit
          in: query
          schema:
            type: number
            default: 20
        - name: search
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
            enum: [active, inactive, trial, subscribed]
      responses:
        "200":
          description: Users list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedAdminUsers"
  /admin/users/{id}:
    get:
      operationId: getAdminUser
      tags: [admin]
      summary: Get user details (admin)
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "200":
          description: User details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminUserDetail"
    put:
      operationId: updateAdminUser
      tags: [admin]
      summary: Update user (admin)
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/IdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminUpdateUserRequest"
      responses:
        "200":
          description: User updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminUserDetail"
  /admin/users/{id}/toggle-status:
    put:
      operationId: toggleUserStatus
      tags: [admin]
      summary: Toggle user active/inactive status
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "200":
          description: Status toggled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /admin/plans:
    get:
      operationId: getAdminPlans
      tags: [admin]
      summary: Get all plans (admin)
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Plans list
          content:
            application/json:
              schema:
                type: object
                required: [plans]
                properties:
                  plans:
                    type: array
                    items:
                      $ref: "#/components/schemas/Plan"
    post:
      operationId: createPlan
      tags: [admin]
      summary: Create subscription plan
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePlanRequest"
      responses:
        "201":
          description: Plan created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Plan"
  /admin/plans/{id}:
    put:
      operationId: updatePlan
      tags: [admin]
      summary: Update subscription plan
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/IdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdatePlanRequest"
      responses:
        "200":
          description: Plan updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Plan"
    delete:
      operationId: deletePlan
      tags: [admin]
      summary: Delete subscription plan
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "200":
          description: Plan deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /admin/currencies:
    get:
      operationId: getAdminCurrencies
      tags: [admin]
      summary: Get all currencies (admin)
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Currencies
          content:
            application/json:
              schema:
                type: object
                required: [currencies]
                properties:
                  currencies:
                    type: array
                    items:
                      $ref: "#/components/schemas/Currency"
    post:
      operationId: addCurrency
      tags: [admin]
      summary: Add a currency
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCurrencyRequest"
      responses:
        "201":
          description: Currency added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Currency"
  /admin/currencies/{id}/exchange-rate:
    put:
      operationId: updateExchangeRate
      tags: [admin]
      summary: Update currency exchange rate
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/IdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateExchangeRateInput"
      responses:
        "200":
          description: Exchange rate updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Currency"
  /admin/subscriptions:
    get:
      operationId: getAdminSubscriptions
      tags: [admin]
      summary: Get all subscriptions (admin)
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: number
            default: 1
        - name: limit
          in: query
          schema:
            type: number
            default: 20
        - name: status
          in: query
          schema:
            type: string
            enum: [active, cancelled, expired, trial]
      responses:
        "200":
          description: Subscriptions list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedAdminSubscriptions"
  /admin/settings:
    get:
      operationId: getAdminSettings
      tags: [admin]
      summary: Get system settings
      security:
        - bearerAuth: []
      responses:
        "200":
          description: System settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SystemSettings"
    put:
      operationId: updateAdminSettings
      tags: [admin]
      summary: Update system settings
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateSystemSettingsRequest"
      responses:
        "200":
          description: Settings updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SystemSettings"
  /admin/send-notification:
    post:
      operationId: sendAdminNotification
      tags: [admin]
      summary: Send push notification to users
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SendNotificationRequest"
      responses:
        "200":
          description: Notification sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

# ─── COMPONENTS ───────────────────────────────────────────────
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    IdParam:
      name: id
      in: path
      required: true
      schema:
        type: string
  schemas:
...
    SuccessResponse:
      type: object
      required: [success, message]
      properties:
        success:
          type: boolean
        message:
          type: string
...
    UpdateExchangeRateInput:
      type: object
      required: [rate]
      properties:
        rate:
          type: number
        base:
          type: string
...
    Currency:
      type: object
      required: [id, code, name, symbol, isActive]
      properties:
        id:
          type: string
        code:
          type: string
          example: USD
        name:
          type: string
        nameAr:
          type: string
          nullable: true
        symbol:
          type: string
        flag:
          type: string
          nullable: true
        rateToUsd:
          type: number
          nullable: true
        isActive:
          type: boolean
        updatedAt:
          type: string
          format: date-time
          nullable: true
...
    CreateCurrencyRequest:
      type: object
      required: [code, name, symbol]
      properties:
        code:
          type: string
        name:
          type: string
        nameAr:
          type: string
        symbol:
          type: string
        flag:
          type: string
        rateToUsd:
          type: number

    # ─── BUDGETS ─────────────────────────────────────────────────
...
    Plan:
      type: object
      required: [id, name, price, currency, interval, transactionLimit, features, isActive]
      properties:
        id:
          type: string
        name:
          type: string
        nameAr:
          type: string
          nullable: true
        description:
          type: string
          nullable: true
        price:
          type: number
        currency:
          type: string
        interval:
          type: string
          enum: [monthly, yearly, lifetime]
        transactionLimit:
          type: number
          nullable: true
          description: null means unlimited
        features:
          type: array
          items:
            type: string
        isActive:
          type: boolean
        isFree:
          type: boolean
        trialDays:
          type: number
          nullable: true
        googlePlayProductId:
          type: string
          nullable: true
        appleProductId:
          type: string
          nullable: true
    UserSubscription:
      type: object
      required: [id, planId, plan, status, startDate, expiresAt]
      properties:
        id:
          type: string
        planId:
          type: string
        plan:
          $ref: "#/components/schemas/Plan"
        status:
          type: string
          enum: [trial, active, cancelled, expired]
        startDate:
          type: string
          format: date-time
        expiresAt:
          type: string
          format: date-time
        cancelledAt:
          type: string
          format: date-time
          nullable: true
        autoRenew:
          type: boolean
...
    CreatePlanRequest:
      type: object
      required: [name, price, currency, interval, features]
      properties:
        name:
          type: string
        nameAr:
          type: string
        description:
          type: string
        price:
          type: number
        currency:
          type: string
        interval:
          type: string
          enum: [monthly, yearly, lifetime]
        transactionLimit:
          type: number
        features:
          type: array
          items:
            type: string
        isFree:
          type: boolean
        trialDays:
          type: number
        googlePlayProductId:
          type: string
        appleProductId:
          type: string
    UpdatePlanRequest:
      type: object
      properties:
        name:
          type: string
        nameAr:
          type: string
        description:
          type: string
        price:
          type: number
        transactionLimit:
          type: number
        features:
          type: array
          items:
            type: string
        isActive:
          type: boolean

    # ─── PAYMENTS ─────────────────────────────────────────────────
...
    SendNotificationRequest:
      type: object
      required: [title, body]
      properties:
        title:
          type: string
        body:
          type: string
        targetUsers:
          type: string
          enum: [all, subscribed, trial, free]
        userIds:
          type: array
          items:
            type: string
        type:
          type: string
          enum: [budget_alert, subscription, system, transaction, backup]

    # ─── BACKUPS ─────────────────────────────────────────────────
...
    UserSettings:
      type: object
      required: [id, userId, theme, language, currencyCode, dateFormat, notificationsEnabled]
      properties:
        id:
          type: string
        userId:
          type: string
        theme:
          type: string
          enum: [light, dark, system]
        language:
          type: string
          enum: [ar, en]
        currencyCode:
          type: string
        dateFormat:
          type: string
        numberFormat:
          type: string
        notificationsEnabled:
          type: boolean
        budgetAlerts:
          type: boolean
        weeklyReport:
          type: boolean
        backupEnabled:
          type: boolean
        backupFrequency:
          type: string
          enum: [daily, weekly, monthly]
    UpdateSettingsRequest:
      type: object
      properties:
        theme:
          type: string
          enum: [light, dark, system]
        language:
          type: string
          enum: [ar, en]
        currencyCode:
          type: string
        dateFormat:
          type: string
        numberFormat:
          type: string
        notificationsEnabled:
          type: boolean
        budgetAlerts:
          type: boolean
        weeklyReport:
          type: boolean
        backupEnabled:
          type: boolean
        backupFrequency:
          type: string
          enum: [daily, weekly, monthly]

    # ─── ADMIN ───────────────────────────────────────────────────
    AdminStats:
      type: object
      required: [totalUsers, activeUsers, trialUsers, subscribedUsers, totalRevenue, monthlyRevenue, totalTransactions, newUsersThisMonth]
      properties:
        totalUsers:
          type: number
        activeUsers:
          type: number
        trialUsers:
          type: number
        subscribedUsers:
          type: number
        totalRevenue:
          type: number
        monthlyRevenue:
          type: number
        totalTransactions:
          type: number
        newUsersThisMonth:
          type: number
        conversionRate:
          type: number
    RevenueData:
      type: object
      required: [labels, revenue, subscriptions]
      properties:
        labels:
          type: array
          items:
            type: string
        revenue:
          type: array
          items:
            type: number
        subscriptions:
          type: array
          items:
            type: number
        total:
          type: number
    AdminUserDetail:
      type: object
      required: [id, email, name, role, isActive, createdAt]
      properties:
        id:
          type: string
        email:
          type: string
        name:
          type: string
        phone:
          type: string
          nullable: true
        country:
          type: string
          nullable: true
        currencyCode:
          type: string
        avatarUrl:
          type: string
          nullable: true
        role:
          type: string
          enum: [user, admin]
        isActive:
          type: boolean
        subscription:
          $ref: "#/components/schemas/UserSubscription"
          nullable: true
        transactionCount:
          type: number
        accountCount:
          type: number
        lastLoginAt:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
    PaginatedAdminUsers:
      type: object
      required: [users, total, page, limit, totalPages]
      properties:
        users:
          type: array
          items:
            $ref: "#/components/schemas/AdminUserDetail"
        total:
          type: number
        page:
          type: number
        limit:
          type: number
        totalPages:
          type: number
    AdminUpdateUserRequest:
      type: object
      properties:
        name:
          type: string
        role:
          type: string
          enum: [user, admin]
        isActive:
          type: boolean
        currencyCode:
          type: string
    PaginatedAdminSubscriptions:
      type: object
      required: [subscriptions, total, page, limit, totalPages]
      properties:
        subscriptions:
          type: array
          items:
            type: object
            required: [id, userId, userName, planName, status, startDate, expiresAt]
            properties:
              id:
                type: string
              userId:
                type: string
              userName:
                type: string
              userEmail:
                type: string
              planName:
                type: string
              status:
                type: string
                enum: [trial, active, cancelled, expired]
              startDate:
                type: string
                format: date-time
              expiresAt:
                type: string
                format: date-time
              amount:
                type: number
        total:
          type: number
        page:
          type: number
        limit:
          type: number
        totalPages:
          type: number
    SystemSettings:
      type: object
      required: [appName, trialDays, freeTransactionLimit, maintenanceMode]
      properties:
        appName:
          type: string
        appNameAr:
          type: string
        supportEmail:
          type: string
        trialDays:
          type: number
        freeTransactionLimit:
          type: number
        maintenanceMode:
          type: boolean
        registrationEnabled:
          type: boolean
        defaultCurrency:
          type: string
        defaultLanguage:
          type: string
          enum: [ar, en]
    UpdateSystemSettingsRequest:
      type: object
      properties:
        appName:
          type: string
        appNameAr:
          type: string
        supportEmail:
          type: string
        trialDays:
          type: number
        freeTransactionLimit:
          type: number
        maintenanceMode:
          type: boolean
        registrationEnabled:
          type: boolean
        defaultCurrency:
          type: string
        defaultLanguage:
          type: string
          enum: [ar, en]